requirements.txt
pytesseract==0.3.10
Pillow==10.2.0
requests==2.31.0
praw==7.7.1
python-dotenv==1.0.1 